// apps/api/src/generated/kysely/knowledge/PropertyValueType.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Represents the enum knowledge.property_value_type */
enum PropertyValueType {
  text = 'text',
  number = 'number',
  date = 'date',
  boolean = 'boolean',
  json = 'json',
};

export default PropertyValueType;